Module aws_sdk_transfer::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_service_error

See InternalServiceError

invalid_next_token_exception

See InvalidNextTokenException

invalid_request_exception

See InvalidRequestException

resource_exists_exception

See ResourceExistsException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

throttling_exception

See ThrottlingException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

ConflictException

This exception is thrown when the UpdateServer is called for a file transfer

CreateAccessError

CreateServerError

CreateUserError

CreateWorkflowError

DeleteAccessError

DeleteServerError

DeleteSshPublicKeyError

DeleteUserError

DeleteWorkflowError

DescribeAccessError

DescribeExecutionError

DescribeSecurityPolicyError

DescribeServerError

DescribeUserError

DescribeWorkflowError

ImportSshPublicKeyError

InternalServiceError

This exception is thrown when an error occurs in the Amazon Web ServicesTransfer Family service.

InvalidNextTokenException

The NextToken parameter that was passed is invalid.

InvalidRequestException

This exception is thrown when the client submits a malformed request.

ListAccessesError

ListExecutionsError

ListSecurityPoliciesError

ListServersError

ListTagsForResourceError

ListUsersError

ListWorkflowsError

ResourceExistsException

The requested resource does not exist.

ResourceNotFoundException

This exception is thrown when a resource is not found by the Amazon Web ServicesTransfer Family

SendWorkflowStepStateError

ServiceUnavailableException

The request has failed because the Amazon Web ServicesTransfer Family service is not available.

StartServerError

StopServerError

TagResourceError

TestIdentityProviderError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateAccessError

UpdateServerError

UpdateUserError

Enums

CreateAccessErrorKind

CreateServerErrorKind

CreateUserErrorKind

CreateWorkflowErrorKind

DeleteAccessErrorKind

DeleteServerErrorKind

DeleteSshPublicKeyErrorKind

DeleteUserErrorKind

DeleteWorkflowErrorKind

DescribeAccessErrorKind

DescribeExecutionErrorKind

DescribeSecurityPolicyErrorKind

DescribeServerErrorKind

DescribeUserErrorKind

DescribeWorkflowErrorKind

ImportSshPublicKeyErrorKind

ListAccessesErrorKind

ListExecutionsErrorKind

ListSecurityPoliciesErrorKind

ListServersErrorKind

ListTagsForResourceErrorKind

ListUsersErrorKind

ListWorkflowsErrorKind

SendWorkflowStepStateErrorKind

StartServerErrorKind

StopServerErrorKind

TagResourceErrorKind

TestIdentityProviderErrorKind

UntagResourceErrorKind

UpdateAccessErrorKind

UpdateServerErrorKind

UpdateUserErrorKind