Struct aws_sdk_transfer::input::create_server_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateServerInput

Implementations

impl Builder

pub fn certificate(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon Web Services Certificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

To request a new public certificate, see Request a public certificate in the Amazon Web Services Certificate Manager User Guide.

To import an existing certificate into ACM, see Importing certificates into ACM in the Amazon Web Services Certificate Manager User Guide.

To request a private certificate to use FTPS through private IP addresses, see Request a private certificate in the Amazon Web Services Certificate Manager User Guide.

Certificates with the following cryptographic algorithms and key sizes are supported:

2048-bit RSA (RSA_2048)

4096-bit RSA (RSA_4096)

Elliptic Prime Curve 256 bit (EC_prime256v1)

Elliptic Prime Curve 384 bit (EC_secp384r1)

Elliptic Prime Curve 521 bit (EC_secp521r1)

The certificate must be a valid SSL/TLS X.509 version 3 certificate with FQDN or IP address specified and information about the issuer.

pub fn set_certificate(self, input: Option<String>) -> Self

pub fn domain(self, input: Domain) -> Self

The domain of the storage system that is used for file transfers. There are two domains available: Amazon Simple Storage Service (Amazon S3) and Amazon Elastic File System (Amazon EFS). The default value is S3.

After the server is created, the domain cannot be changed.

pub fn set_domain(self, input: Option<Domain>) -> Self

pub fn endpoint_details(self, input: EndpointDetails) -> Self

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

pub fn set_endpoint_details(self, input: Option<EndpointDetails>) -> Self

pub fn endpoint_type(self, input: EndpointType) -> Self

The type of endpoint that you want your server to use. You can choose to make your server's endpoint publicly accessible (PUBLIC) or host it inside your VPC. With an endpoint that is hosted in a VPC, you can restrict access to your server and resources only within your VPC or choose to make it internet facing by attaching Elastic IP addresses directly to it.

After May 19, 2021, you won't be able to create a server using EndpointType=VPC_ENDPOINT in your Amazon Web Services account if your account hasn't already done so before May 19, 2021. If you have already created servers with EndpointType=VPC_ENDPOINT in your Amazon Web Services account on or before May 19, 2021, you will not be affected. After this date, use EndpointType=VPC.

For more information, see https://docs.aws.amazon.com/transfer/latest/userguide/create-server-in-vpc.html#deprecate-vpc-endpoint.

It is recommended that you use VPC as the EndpointType. With this endpoint type, you have the option to directly associate up to three Elastic IPv4 addresses (BYO IP included) with your server's endpoint and use VPC security groups to restrict traffic by the client's public IP address. This is not possible with EndpointType set to VPC_ENDPOINT.

pub fn set_endpoint_type(self, input: Option<EndpointType>) -> Self

pub fn host_key(self, input: impl Into<String>) -> Self

The RSA private key as generated by the ssh-keygen -N "" -m PEM -f my-new-server-key command.

If you aren't planning to migrate existing users from an existing SFTP-enabled server to a new server, don't update the host key. Accidentally changing a server's host key can be disruptive.

For more information, see Change the host key for your SFTP-enabled server in the Amazon Web Services Transfer Family User Guide.

pub fn set_host_key(self, input: Option<String>) -> Self

pub fn identity_provider_details(self, input: IdentityProviderDetails) -> Self

Required when IdentityProviderType is set to AWS_DIRECTORY_SERVICE or API_GATEWAY. Accepts an array containing all of the information required to use a directory in AWS_DIRECTORY_SERVICE or invoke a customer-supplied authentication API, including the API Gateway URL. Not required when IdentityProviderType is set to SERVICE_MANAGED.

pub fn set_identity_provider_details(
self,
input: Option<IdentityProviderDetails>
) -> Self

pub fn identity_provider_type(self, input: IdentityProviderType) -> Self

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

pub fn set_identity_provider_type(
self,
input: Option<IdentityProviderType>
) -> Self

pub fn logging_role(self, input: impl Into<String>) -> Self

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

pub fn set_logging_role(self, input: Option<String>) -> Self

pub fn protocols(self, input: impl Into<Protocol>) -> Self

pub fn set_protocols(self, input: Option<Vec<Protocol>>) -> Self

pub fn security_policy_name(self, input: impl Into<String>) -> Self

Specifies the name of the security policy that is attached to the server.

pub fn set_security_policy_name(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn workflow_details(self, input: WorkflowDetails) -> Self

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

pub fn set_workflow_details(self, input: Option<WorkflowDetails>) -> Self

pub fn build(self) -> Result<CreateServerInput, BuildError>

Consumes the builder and constructs a CreateServerInput