Module aws_sdk_transfer::model

Modules

copy_step_details

See CopyStepDetails

custom_step_details

See CustomStepDetails

delete_step_details

See DeleteStepDetails

described_access

See DescribedAccess

described_execution

See DescribedExecution

described_security_policy

See DescribedSecurityPolicy

described_server

See DescribedServer

described_user

See DescribedUser

described_workflow

See DescribedWorkflow

efs_file_location

See EfsFileLocation

endpoint_details

See EndpointDetails

execution_error

See ExecutionError

execution_results

See ExecutionResults

execution_step_result

See ExecutionStepResult

file_location

See FileLocation

home_directory_map_entry

See HomeDirectoryMapEntry

identity_provider_details

See IdentityProviderDetails

input_file_location

See InputFileLocation

listed_access

See ListedAccess

listed_execution

See ListedExecution

listed_server

See ListedServer

listed_user

See ListedUser

listed_workflow

See ListedWorkflow

logging_configuration

See LoggingConfiguration

posix_profile

See PosixProfile

protocol_details

See ProtocolDetails

s3_file_location

See S3FileLocation

s3_input_file_location

See S3InputFileLocation

s3_tag

See S3Tag

service_metadata

See ServiceMetadata

ssh_public_key

See SshPublicKey

tag

See Tag

tag_step_details

See TagStepDetails

user_details

See UserDetails

workflow_detail

See WorkflowDetail

workflow_details

See WorkflowDetails

workflow_step

See WorkflowStep

Structs

CopyStepDetails

Each step type has its own StepDetails structure.

CustomStepDetails

Each step type has its own StepDetails structure.

DeleteStepDetails

The name of the step, used to identify the step that is being deleted.

DescribedAccess

Describes the properties of the access that was specified.

DescribedExecution

The details for an execution object.

DescribedSecurityPolicy

Describes the properties of a security policy that was specified. For more information

DescribedServer

Describes the properties of a file transfer protocol-enabled server that was

DescribedUser

Describes the properties of a user that was specified.

DescribedWorkflow

Describes the properties of the specified workflow

EfsFileLocation

Specifies the details for the file location for the file being used in the workflow. Only applicable if you are using Amazon EFS for storage.

EndpointDetails

The virtual private cloud (VPC) endpoint settings that are configured for your file

ExecutionError

Specifies the error message and type, for an error that occurs during the execution of the workflow.

ExecutionResults

Specifies the steps in the workflow, as well as the steps to execute in case of any errors during workflow execution.

ExecutionStepResult

Specifies the following details for the step: error (if any), outputs (if any), and the step type.

FileLocation

Specifies the Amazon S3 or EFS file details to be used in the step.

HomeDirectoryMapEntry

Represents an object that contains entries and targets for

IdentityProviderDetails

Returns information related to the type of user authentication that is in use for a file

InputFileLocation

Specifies the location for the file being copied. Only applicable for the Copy type of workflow steps.

ListedAccess

Lists the properties for one or more specified associated accesses.

ListedExecution

Returns properties of the execution that is specified.

ListedServer

Returns properties of a file transfer protocol-enabled server that was specified.

ListedUser

Returns properties of the user that you specify.

ListedWorkflow

Contains the ID, text description, and Amazon Resource Name (ARN) for the workflow.

LoggingConfiguration

Consists of the logging role and the log group name.

PosixProfile

The full POSIX identity, including user ID (Uid), group ID

ProtocolDetails

S3FileLocation

Specifies the details for the file location for the file being used in the workflow. Only applicable if you are using S3 storage.

S3InputFileLocation

Specifies the details for the S3 file being copied.

S3Tag

Specifies the key-value pair that are assigned to a file during the execution of a Tagging step.

ServiceMetadata

A container object for the session details associated with a workflow.

SshPublicKey

Provides information about the public Secure Shell (SSH) key that is associated with a

Tag

Creates a key-value pair for a specific resource. Tags are metadata that you can use to

TagStepDetails

Each step type has its own StepDetails structure.

UserDetails

Specifies the user name, server ID, and session ID for a workflow.

WorkflowDetail

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

WorkflowDetails

Container for the WorkflowDetail data type.

WorkflowStep

The basic building block of a workflow.

Enums

CustomStepStatus

Domain

EndpointType

ExecutionErrorType

ExecutionStatus

HomeDirectoryType

IdentityProviderType

Returns information related to the type of user authentication that is in use for a file

OverwriteExisting

Protocol

State

Describes the condition of a file transfer protocol-enabled server with respect to its

WorkflowStepType