Struct aws_sdk_transfer::model::described_server::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribedServer

Implementations

impl Builder

pub fn arn(self, input: impl Into<String>) -> Self

Specifies the unique Amazon Resource Name (ARN) of the server.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn certificate(self, input: impl Into<String>) -> Self

Specifies the ARN of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

pub fn set_certificate(self, input: Option<String>) -> Self

pub fn protocol_details(self, input: ProtocolDetails) -> Self

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode. Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

pub fn set_protocol_details(self, input: Option<ProtocolDetails>) -> Self

pub fn domain(self, input: Domain) -> Self

Specifies the domain of the storage system that is used for file transfers.

pub fn set_domain(self, input: Option<Domain>) -> Self

pub fn endpoint_details(self, input: EndpointDetails) -> Self

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

pub fn set_endpoint_details(self, input: Option<EndpointDetails>) -> Self

pub fn endpoint_type(self, input: EndpointType) -> Self

Defines the type of endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

pub fn set_endpoint_type(self, input: Option<EndpointType>) -> Self

pub fn host_key_fingerprint(self, input: impl Into<String>) -> Self

Specifies the Base64-encoded SHA256 fingerprint of the server's host key. This value is equivalent to the output of the ssh-keygen -l -f my-new-server-key command.

pub fn set_host_key_fingerprint(self, input: Option<String>) -> Self

pub fn identity_provider_details(self, input: IdentityProviderDetails) -> Self

Specifies information to call a customer-supplied authentication API. This field is not populated when the IdentityProviderType of a server is AWS_DIRECTORY_SERVICE or SERVICE_MANAGED.

pub fn set_identity_provider_details(
self,
input: Option<IdentityProviderDetails>
) -> Self

pub fn identity_provider_type(self, input: IdentityProviderType) -> Self

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

pub fn set_identity_provider_type(
self,
input: Option<IdentityProviderType>
) -> Self

pub fn logging_role(self, input: impl Into<String>) -> Self

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

pub fn set_logging_role(self, input: Option<String>) -> Self

pub fn protocols(self, input: impl Into<Protocol>) -> Self

pub fn set_protocols(self, input: Option<Vec<Protocol>>) -> Self

pub fn security_policy_name(self, input: impl Into<String>) -> Self

Specifies the name of the security policy that is attached to the server.

pub fn set_security_policy_name(self, input: Option<String>) -> Self

pub fn server_id(self, input: impl Into<String>) -> Self

Specifies the unique system-assigned identifier for a server that you instantiate.

pub fn set_server_id(self, input: Option<String>) -> Self

pub fn state(self, input: State) -> Self

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

pub fn set_state(self, input: Option<State>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn user_count(self, input: i32) -> Self

Specifies the number of users that are assigned to a server you specified with the ServerId.

pub fn set_user_count(self, input: Option<i32>) -> Self

pub fn workflow_details(self, input: WorkflowDetails) -> Self

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

pub fn set_workflow_details(self, input: Option<WorkflowDetails>) -> Self

pub fn build(self) -> DescribedServer

Consumes the builder and constructs a DescribedServer