Struct aws_sdk_transfer::model::DescribedServer

#[non_exhaustive]
pub struct DescribedServer {

    pub arn: Option<String>,
    pub certificate: Option<String>,
    pub protocol_details: Option<ProtocolDetails>,
    pub domain: Option<Domain>,
    pub endpoint_details: Option<EndpointDetails>,
    pub endpoint_type: Option<EndpointType>,
    pub host_key_fingerprint: Option<String>,
    pub identity_provider_details: Option<IdentityProviderDetails>,
    pub identity_provider_type: Option<IdentityProviderType>,
    pub logging_role: Option<String>,
    pub protocols: Option<Vec<Protocol>>,
    pub security_policy_name: Option<String>,
    pub server_id: Option<String>,
    pub state: Option<State>,
    pub tags: Option<Vec<Tag>>,
    pub user_count: Option<i32>,
    pub workflow_details: Option<WorkflowDetails>,

}

Describes the properties of a file transfer protocol-enabled server that was specified.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

arn: Option<String>

Specifies the unique Amazon Resource Name (ARN) of the server.

certificate: Option<String>

Specifies the ARN of the Amazon Web ServicesCertificate Manager (ACM) certificate. Required when Protocols is set to FTPS.

protocol_details: Option<ProtocolDetails>

The protocol settings that are configured for your server.

Use the PassiveIp parameter to indicate passive mode. Enter a single dotted-quad IPv4 address, such as the external IP address of a firewall, router, or load balancer.

domain: Option<Domain>

Specifies the domain of the storage system that is used for file transfers.

endpoint_details: Option<EndpointDetails>

The virtual private cloud (VPC) endpoint settings that are configured for your server. When you host your endpoint within your VPC, you can make it accessible only to resources within your VPC, or you can attach Elastic IP addresses and make it accessible to clients over the internet. Your VPC's default security groups are automatically assigned to your endpoint.

endpoint_type: Option<EndpointType>

Defines the type of endpoint that your server is connected to. If your server is connected to a VPC endpoint, your server isn't accessible over the public internet.

host_key_fingerprint: Option<String>

Specifies the Base64-encoded SHA256 fingerprint of the server's host key. This value is equivalent to the output of the ssh-keygen -l -f my-new-server-key command.

identity_provider_details: Option<IdentityProviderDetails>

Specifies information to call a customer-supplied authentication API. This field is not populated when the IdentityProviderType of a server is AWS_DIRECTORY_SERVICE or SERVICE_MANAGED.

identity_provider_type: Option<IdentityProviderType>

Specifies the mode of authentication for a server. The default value is SERVICE_MANAGED, which allows you to store and access user credentials within the Amazon Web Services Transfer Family service.

Use AWS_DIRECTORY_SERVICE to provide access to Active Directory groups in Amazon Web Services Managed Active Directory or Microsoft Active Directory in your on-premises environment or in Amazon Web Services using AD Connectors. This option also requires you to provide a Directory ID using the IdentityProviderDetails parameter.

Use the API_GATEWAY value to integrate with an identity provider of your choosing. The API_GATEWAY setting requires you to provide an API Gateway endpoint URL to call for authentication using the IdentityProviderDetails parameter.

logging_role: Option<String>

Specifies the Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that allows a server to turn on Amazon CloudWatch logging for Amazon S3 or Amazon EFS events. When set, user activity can be viewed in your CloudWatch logs.

protocols: Option<Vec<Protocol>>

Specifies the file transfer protocol or protocols over which your file transfer protocol client can connect to your server's endpoint. The available protocols are:

SFTP (Secure Shell (SSH) File Transfer Protocol): File transfer over SSH

FTPS (File Transfer Protocol Secure): File transfer with TLS encryption

FTP (File Transfer Protocol): Unencrypted file transfer

security_policy_name: Option<String>

Specifies the name of the security policy that is attached to the server.

server_id: Option<String>

Specifies the unique system-assigned identifier for a server that you instantiate.

state: Option<State>

Specifies the condition of a server for the server that was described. A value of ONLINE indicates that the server can accept jobs and transfer files. A State value of OFFLINE means that the server cannot perform file transfer operations.

The states of STARTING and STOPPING indicate that the server is in an intermediate state, either not fully able to respond, or not fully offline. The values of START_FAILED or STOP_FAILED can indicate an error condition.

tags: Option<Vec<Tag>>

Specifies the key-value pairs that you can use to search for and group servers that were assigned to the server that was described.

user_count: Option<i32>

Specifies the number of users that are assigned to a server you specified with the ServerId.

workflow_details: Option<WorkflowDetails>

Specifies the workflow ID for the workflow to assign and the execution role used for executing the workflow.

Implementations

impl DescribedServer

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribedServer