Module aws_sdk_transfer::operation

Structs

CreateAccess

Used by administrators to choose which groups in the directory should have access to

CreateServer

Instantiates an auto-scaling virtual server based on the selected file transfer protocol

CreateUser

Creates a user and associates them with an existing file transfer protocol-enabled server.

CreateWorkflow

DeleteAccess

Allows you to delete the access specified in the ServerID and

DeleteServer

Deletes the file transfer protocol-enabled server that you specify.

DeleteSshPublicKey

Deletes a user's Secure Shell (SSH) public key.

DeleteUser

Deletes the user belonging to a file transfer protocol-enabled server you specify.

DeleteWorkflow

Deletes the specified workflow.

DescribeAccess

Describes the access that is assigned to the specific file transfer protocol-enabled

DescribeExecution

You can use DescribeExecution to check the details of the execution of the specified workflow.

DescribeSecurityPolicy

Describes the security policy that is attached to your file transfer protocol-enabled

DescribeServer

Describes a file transfer protocol-enabled server that you specify by passing the

DescribeUser

Describes the user assigned to the specific file transfer protocol-enabled server, as

DescribeWorkflow

Describes the specified workflow.

ImportSshPublicKey

Adds a Secure Shell (SSH) public key to a user account identified by a

ListAccesses

Lists the details for all the accesses you have on your server.

ListExecutions

Lists all executions for the specified workflow.

ListSecurityPolicies

Lists the security policies that are attached to your file transfer protocol-enabled

ListServers

Lists the file transfer protocol-enabled servers that are associated with your Amazon Web Services

ListTagsForResource

Lists all of the tags associated with the Amazon Resource Name (ARN) that you specify. The

ListUsers

Lists the users for a file transfer protocol-enabled server that you specify by passing

ListWorkflows

Lists all of your workflows.

SendWorkflowStepState

Sends a callback for asynchronous custom steps.

StartServer

Changes the state of a file transfer protocol-enabled server from OFFLINE to

StopServer

Changes the state of a file transfer protocol-enabled server from ONLINE to

TagResource

Attaches a key-value pair to a resource, as identified by its Amazon Resource Name (ARN).

TestIdentityProvider

If the IdentityProviderType of a file transfer protocol-enabled server is

UntagResource

Detaches a key-value pair from a resource, as identified by its Amazon Resource Name

UpdateAccess

Allows you to update parameters for the access specified in the ServerID and

UpdateServer

Updates the file transfer protocol-enabled server's properties after that server has

UpdateUser

Assigns new properties to a user. Parameters you pass modify any or all of the following: