Module aws_sdk_transfer::model

Modules

described_access

See DescribedAccess

described_security_policy

See DescribedSecurityPolicy

described_server

See DescribedServer

described_user

See DescribedUser

endpoint_details

See EndpointDetails

home_directory_map_entry

See HomeDirectoryMapEntry

identity_provider_details

See IdentityProviderDetails

listed_access

See ListedAccess

listed_server

See ListedServer

listed_user

See ListedUser

posix_profile

See PosixProfile

protocol_details

See ProtocolDetails

ssh_public_key

See SshPublicKey

tag

See Tag

Structs

DescribedAccess

Describes the properties of the access that was specified.

DescribedSecurityPolicy

Describes the properties of a security policy that was specified. For more information

DescribedServer

Describes the properties of a file transfer protocol-enabled server that was

DescribedUser

Describes the properties of a user that was specified.

EndpointDetails

The virtual private cloud (VPC) endpoint settings that are configured for your file

HomeDirectoryMapEntry

Represents an object that contains entries and targets for

IdentityProviderDetails

Returns information related to the type of user authentication that is in use for a file

ListedAccess

Lists the properties for one or more specified associated accesses.

ListedServer

Returns properties of a file transfer protocol-enabled server that was specified.

ListedUser

Returns properties of the user that you specify.

PosixProfile

The full POSIX identity, including user ID (Uid), group ID

ProtocolDetails

SshPublicKey

Provides information about the public Secure Shell (SSH) key that is associated with a

Tag

Creates a key-value pair for a specific resource. Tags are metadata that you can use to

Enums

Domain

EndpointType

HomeDirectoryType

IdentityProviderType

Returns information related to the type of user authentication that is in use for a file

Protocol

State

Describes the condition of a file transfer protocol-enabled server with respect to its